Move ClusterPersistRepositoryConfigurationYamlSwapper to mode module
