+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/yaml/YamlScalingConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.yaml.pojo.scaling;
+package org.apache.shardingsphere.governance.core.yaml;
 
 import lombok.Getter;
 import lombok.Setter;
