+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.yaml.config.swapper.mode.PersistRepositoryConfigurationYamlSwapper
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.governance.core.yaml.swapper.ClusterPersistRepositoryConfigurationYamlSwapper
+org.apache.shardingsphere.mode.manager.cluster.yaml.ClusterPersistRepositoryConfigurationYamlSwapper
